refactor(tint): split god-component TintImport (726→387)
Quebra a página de Importação Tintométrica em módulos sob src/components/tintImport/.
Os handlers de import (chunking/retry/resume, ~340L de orquestração delicada)
permanecem VERBATIM no pai; o resto vira módulo + componentes controlados.

- types.ts: constantes (ACCOUNT/CHUNK_SIZE_*/MAX_RETRIES) + tipos + helpers puros
  (getChunkSize/sha256/sleep/sendChunkWithRetry) + TIPO_OPTIONS + statusColor.
- queries.ts: useImportHistory + useTintProductCounts.
- SyncCard.tsx: card de sincronização Omie.
- ImportCard.tsx: tipo/modo + dropzone + preview + progresso + resultados + botão.
- HistoryTable.tsx: histórico com botão Retomar.
- +14 testes (4 arquivos): helpers (getChunkSize/sha256/sendChunkWithRetry), SyncCard,
  ImportCard (preview/resultados/disabled) e HistoryTable (Retomar).

tsc --noEmit limpo; lint 0 problemas.
